[PATCH 2/2] DRM: Add support for drmModeAddFB2
When working in 64-bit, handles passed to the kernel
DRM can not be user-space pointer anymore, but rather
should be a real kernel handles (fds). A change is needed
in the UI components and the kernel to fix this
functionality. in addition, since the multi-plane buffers
need multiple FDs, drmModeAddFB2 is needed to support that.
This is a new API that does not affect current functionality.

Change-Id: Iea5f57fccc43af5e8371cbd31cfd81a35959fe04
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-7099
Signed-off-by: AtanasX D Zlatinski <atanasx.d.zlatinski@intel.com>
Reviewed-on: https://android.intel.com/160389
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Agarwal, Tarun <tarun.agarwal@intel.com>
Tested-by: Agarwal, Tarun <tarun.agarwal@intel.com>
Reviewed-on: https://android.intel.com:443/291373